Changed indirect tower to rotate towards enemies, and working on AOE tower

+++ GGJ2025/GGJ2025/Assets/Scripts/Util/EnemyWaveManager.cs
@@ -53,8 +53,9 @@ public class EnemyWaveManager : MonoBehaviour
 
     public void Update()
     {
+        // Ill move this to HUD manager later
         currenTime -= Time.deltaTime;
-        TimerText.text = "Timer until next wave: "+currenTime;
+        TimerText.text = "Timer until next wave: " + (int)currenTime;
     }
 
     /// <summary>
